Change torch_dtype from float32 to bfloat16

+++ text_to_video/wan2.2-t2v-14b/run_mlperf.py
@@ -67,7 +67,7 @@ class Model:
         self.vae = AutoencoderKLWan.from_pretrained(
             model_path,
             subfolder="vae",
-            torch_dtype=torch.float32
+            torch_dtype=torch.bfloat16
         )
         self.pipe = WanPipeline.from_pretrained(
             model_path,
